chore: rename custom-type.ts to custom.types.ts and update imports
