+++ packages/fsm-core-db-ts/src/custom.types.ts
